F-115: NiA dogfood Proto DataStore + libraryTarget Path B helper
Add libraryTarget for Path B JVM library kinds (protobufLibrary on external
spike). Migrate dogfood UserData from Preferences to typed Proto DataStore;
document F25/F26 and phase C progress.
