+++ plugins/android/src/main/java/library.kt
@@ -1,22 +1,15 @@
-import tools.forma.android.feature.applyFeatures
-import tools.forma.android.feature.kotlinFeatureDefinition
-import tools.forma.android.target.AndroidTargetRegistry
 import tools.forma.android.target.AndroidTargetTypes
 import tools.forma.owners.NoOwner
 import tools.forma.owners.Owner
 import tools.forma.android.visibility.Public
 import tools.forma.android.visibility.Visibility
 import org.gradle.api.Project
-import tools.forma.android.feature.processorConfigurationFeatures
-import tools.forma.deps.core.applyDependencies
-import tools.forma.deps.core.applyTargetPlugins
 import tools.forma.deps.core.FormaDependency
-import tools.forma.deps.fleet.registerFormaLayout
-import tools.forma.validation.asValidator
-import tools.forma.validation.validate
 
 /**
  * Can't depend on api\impl
+ *
+ * Path B derived JVM library kinds (Protobuf, etc.) use [libraryTarget] with a derived type.
  */
 fun Project.library(
     packageName: String,
@@ -25,19 +18,12 @@ fun Project.library(
     visibility: Visibility = Public,
     testDependencies: FormaDependency = emptyDependency()
 ) {
-    AndroidTargetRegistry.selfValidator(AndroidTargetTypes.jvmLibrary).asValidator().validate(target)
-    registerFormaLayout(packageName)
-
-    applyFeatures(
-        kotlinFeatureDefinition()
-    )
-    val processors = processorConfigurationFeatures()
-    applyTargetPlugins(AndroidTargetTypes.jvmLibrary, configurationFeatures = processors)
-
-    applyDependencies(
-        validator = AndroidTargetRegistry.validatorFor(AndroidTargetTypes.jvmLibrary).asValidator(),
+    libraryTarget(
+        type = AndroidTargetTypes.jvmLibrary,
+        packageName = packageName,
         dependencies = dependencies,
+        owner = owner,
+        visibility = visibility,
         testDependencies = testDependencies,
-        configurationFeatures = processors
     )
 }
